+++ operations/eu-central-1/env-eks/main.tf
@@ -1,5 +1,5 @@
 module "operations" {
-  source = "github.com/kentrikos/terraform-aws-account-operations?ref=0.4.3"
+  source = "github.com/kentrikos/terraform-aws-account-operations?ref=0.5.0"
 
   product_domain_name = "${var.product_domain_name}"
   environment_type    = "${var.environment_type}"
@@ -26,4 +26,11 @@ module "operations" {
 
   k8s_masters_iam_policies_arns = "${var.k8s_masters_iam_policies_arns}"
   k8s_nodes_iam_policies_arns   = "${var.k8s_nodes_iam_policies_arns}"
+
+  map_roles          = "${var.k8s_map_roles}"
+  map_roles_count    = "${var.k8s_map_roles_count}"
+  map_users          = "${var.k8s_map_users}"
+  map_users_count    = "${var.k8s_map_users_count}"
+  map_accounts       = "${var.k8s_map_accounts}"
+  map_accounts_count = "${var.k8s_map_accounts_count}"
 }