+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Bump application account module to 0.6.0
+- Bump operations account module to 0.5.0
+
+### Added
+- Added configs for mapping users/roles/accounts to IAM access to k8s
 
 ## [0.9.0] - 2019-06-14
 ### Added